R 3.1.2 update compatibility
Added preprocessCore dependency to install instructions. It was
apparently lost or moved during update.
PCoA was broken by changes to latest version of vegan. I think I’ve
fixed this, but code in general should be reworked for better
compatibility with vegan functions. It is not clear if this update will
break PCoA for people with older vegan versions.